# Makes `writeValueRecursively` iterative