@jobarr-amzn 2021-06-16
Thank you for the detailed comments!